Fixed Combine Attribute Naming !MIGHT BREAK PREVIOUS SETUPS!
!THIS UPDATE MIGHT BREAK PREVIOUS SETUPS!
To get those working with the new update make sure to update all your assembler, prim (volume and mesh prim also) and properties nodes in your graph to the most recent version. You can do so in the asset definition toolbar. If you don't see that toolbar on the top of the parameter window
you can enable it under Top Menu Bar/Assets/Asset Definition Toolbar/Show Always

---

The combine attribute naming was not ideal, using the N attribute to store the count for columns/steps, conflicting with the @N vector normal Attribute.
This was now fixed by renaming the combine attributes to a more readable attribute, (combinerad, combinecount etc)